% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{draw_color_circle}
\alias{draw_color_circle}
\title{create a circle of colours}
\usage{
draw_color_circle(colors, n, lwd = 10)
}
\arguments{
\item{colors}{to be plotted in a circle}

\item{n}{the amount of random colors to choose (will clash with colors)}

\item{lwd}{width of the circle line to draw}
}
\value{
the colors, that were randomly selected if n is set and no colors
}
\description{
create a circle of colours
}
\examples{
#circle with 5 random colors
draw_color_circle(n=5)
}
