% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanRegionPlotting.R
\name{drawXML}
\alias{drawXML}
\title{Plot XMLObject}
\usage{
drawXML(mappingXML, return = FALSE)
}
\arguments{
\item{mappingXML}{the XML with the specs to plot}

\item{return}{where to plot (\code{FALSE}) or return (\code{TRUE}) the object}
}
\value{
the object if return is set to \code{TRUE}
}
\description{
This function converts the XML into a plotable grid object
}
\examples{
#draw the human anatomy
drawXML(loadXML())

}
\seealso{
\code{\link{loadXML}} to get the XML object and \code{\link{changePlottingStyle}} to change the drawing style of organs
}
