% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfparsing.R
\name{getStructuralVariants}
\alias{getStructuralVariants}
\title{Reads structural variants vcf}
\usage{
getStructuralVariants(file, minlength = 10000, chrs = c(1:22, "X",
  "Y"), gns, biomaRtDB)
}
\arguments{
\item{file}{the vcf file to be read in}

\item{minlength}{minimum length of a variant to be considered}

\item{chrs}{which chromosomes to restrict the variants on}

\item{gns}{which genes to use to annotate}

\item{biomaRtDB}{which biomart db to use}
}
\value{
data.table with annotated structural variants
}
\description{
Turns the vcf into an annotated data.table
}
\examples{
#minimal example here
}
