% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanRegionPlotting.R
\name{loadXML}
\alias{loadXML}
\title{Create XML object from file}
\usage{
loadXML(file = system.file("extdata", "human_anatomy.ps.xml", package =
  "SHollizeckPackage"))
}
\arguments{
\item{file}{to be converted into an object}
}
\value{
the XML tree object
}
\description{
This is mainly used to load the XML representing the human anatomy for plotting purposes
can be used for other XMLs as well
}
\examples{
#loading the internal xml for plotting
mappingXML <- loadXML()

}
\seealso{
\code{\link{drawXML}} to draw the XML object and \code{\link{changePlottingStyle}} to change the drawing style of organs
}
