% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSampleAnalysis.R
\name{plotSequenzaCellularity}
\alias{plotSequenzaCellularity}
\title{Make a nice plot from cellularities of sequenza}
\usage{
plotSequenzaCellularity(..., inList = NULL, col, labels)
}
\arguments{
\item{...}{allows arbitrary amounts of ci objects}

\item{inList}{ci objects can also be specified in a list object}

\item{col}{the colors to be used for plotting}

\item{labels}{the names of the samples to be used for plotting}
}
\value{
nothing
}
\description{
Make a nice plot from cellularities of sequenza
}
\examples{
#coming soon to your town
}
