% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{random_colors}
\alias{random_colors}
\title{Generate random colours}
\usage{
random_colors(n = stop("No number of colours specified"), seed)
}
\arguments{
\item{n}{amount of colours to sample}

\item{seed}{seed to use for random generation}
}
\value{
List of colors (hex code)
}
\description{
Generate random colours
}
\examples{
#one random colour
random_colors(n=1)

#setting the seed for reproducibility
random_colors(n=1, seed=1234)
}
