% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHollizeckPackage-package.R
\docType{package}
\name{SHollizeckPackage-package}
\alias{SHollizeckPackage}
\alias{SHollizeckPackage-package}
\title{SHollizeckPackage: Private Function Package}
\description{
Contains the functions written during my PhD. Will probably not be useful for anyone else.
}
\author{
\strong{Maintainer}: Sebastian Hollizeck \email{sebastian.hollizeck@patermac.com} (https://orcid.org/0000-0002-9504-3497)

}
\keyword{internal}
