% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanRegionPlotting.R
\name{changePlottingStyle}
\alias{changePlottingStyle}
\title{Change values for plotting}
\usage{
changePlottingStyle(mappingXML, organ, style = "fill", color = "black",
  lwd = "5")
}
\arguments{
\item{mappingXML}{xml representing the picture}

\item{organ}{which organ to change}

\item{style}{which style (fill/stroke)}

\item{color}{which color to plot this organ in}

\item{lwd}{line width to use for stroke style}
}
\value{
TRUE if something got change FALSE otherwise
}
\description{
Change values for plotting
}
\examples{
#changes the left lung to be filled with red color
mappingXML <- loadXML()
changePlottingStyle(mappingXML, "left_lung", color="red")
drawXML(mappingXML)

#reset page
grid::grid.newpage()

#changes the left lung to be outlined with blue color
mappingXML <- loadXML()
changePlottingStyle(mappingXML, "left_lung", color="blue", style='stroke')
drawXML(mappingXML)

}
\seealso{
* \code{\link{loadXML}} to get the object
* \code{\link{getAllOrgans}} to find all available sites
* \code{\link{drawXML}} to draw the xml object
}
