% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanRegionPlotting.R
\name{getAllOrgans}
\alias{getAllOrgans}
\title{Get all available organs}
\usage{
getAllOrgans(mappingXML)
}
\arguments{
\item{mappingXML}{the xml representing the picture}
}
\value{
list of all organs to chose from
}
\description{
this extracts the top level ids of each child of the root xml
}
